protocol: simplify GPIO mapping logic, saving space and allowing more GPIOs
The GPIO mapping logic is replaced with a PROGMEM table (18 bytes long), and
now supports up to 18 GPIO lines on ATmega8.

The protocol and Linux kernel driver are generalized to support up to 255 GPIO
lines, defined at firmware compile time.  The count is probed by the driver
using BOARD_INIT.

This simplification and extension saves 186 bytes of Flash.