firmware: correct & clean up Makefile comments about fuse selection
Fix the comments for ATmega8 to match the actual fuse selection in the
Makefile.  NOTE: the actual fuse selection in the Makefile is wholly unsuitable
(RC oscillator @ 1 MHz, etc.), but that's a different problem.

Remove guidance on other MCUs, because I can't trust that any thought was given
to the subject, and the project is not portable to other MCUs.